+++ src/main/java/com/programmers/springbootbasic/domain/customer/Repository/CustomerRepository.java
@@ -1,4 +1,6 @@
-package com.programmers.springbootbasic.domain.customer;
+package com.programmers.springbootbasic.domain.customer.Repository;
+
+import com.programmers.springbootbasic.domain.customer.Customer;
 
 import java.util.Optional;
