+++ src/main/java/com/programmers/springbootbasic/domain/voucher/Repository/VoucherRepository.java
@@ -1,4 +1,6 @@
-package com.programmers.springbootbasic.domain.voucher;
+package com.programmers.springbootbasic.domain.voucher.Repository;
+
+import com.programmers.springbootbasic.domain.voucher.Voucher;
 
 import java.util.List;
 import java.util.Optional;
