+++ src/main/java/com/programmers/springbootbasic/presentation/controller/VoucherController.java
@@ -2,8 +2,8 @@ package com.programmers.springbootbasic.presentation.controller;
 
 import com.programmers.springbootbasic.common.util.Validator;
 import com.programmers.springbootbasic.domain.voucher.VoucherType;
+import com.programmers.springbootbasic.presentation.controller.dto.Voucher.VoucherCreationRequest;
 import com.programmers.springbootbasic.service.VoucherService;
-import com.programmers.springbootbasic.service.dto.Voucher.VoucherCreationRequest;
 import com.programmers.springbootbasic.service.dto.Voucher.VoucherResponse;
 import com.programmers.springbootbasic.service.dto.Voucher.VoucherResponses;
 import org.springframework.stereotype.Controller;
@@ -22,13 +22,12 @@ public class VoucherController {
         return voucherService.createVoucher(request);
     }
 
-    public VoucherResponses listVouchers() {
-        return voucherService.list();
-    }
-
-
     private void checkValid(String voucherType, Integer amountOrPercent) {
         Validator.checkNullOrBlank(voucherType);
         Validator.checkNullNumber(amountOrPercent);
+    }
+
+    public VoucherResponses listVouchers() {
+        return voucherService.list();
     }
 }
