+++ src/main/java/com/programmers/springbootbasic/presentation/controller/dto/Voucher/VoucherCreationRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.springbootbasic.service.dto.Voucher;
+package com.programmers.springbootbasic.presentation.controller.dto.Voucher;
 
 import com.programmers.springbootbasic.domain.voucher.VoucherType;
