+++ src/main/java/com/programmers/springbootbasic/service/CustomerMapper.java
@@ -1,9 +1,9 @@
 package com.programmers.springbootbasic.service;
 
 import com.programmers.springbootbasic.domain.customer.Customer;
-import com.programmers.springbootbasic.service.dto.Customer.CustomerCreationRequest;
+import com.programmers.springbootbasic.presentation.controller.dto.Customer.CustomerCreationRequest;
+import com.programmers.springbootbasic.presentation.controller.dto.Customer.CustomerUpdateRequest;
 import com.programmers.springbootbasic.service.dto.Customer.CustomerResponse;
-import com.programmers.springbootbasic.service.dto.Customer.CustomerUpdateRequest;
 
 import java.util.UUID;
