+++ src/main/java/com/programmers/springbootbasic/service/CustomerService.java
@@ -1,10 +1,10 @@
 package com.programmers.springbootbasic.service;
 
 import com.programmers.springbootbasic.domain.customer.Customer;
-import com.programmers.springbootbasic.domain.customer.CustomerRepository;
-import com.programmers.springbootbasic.service.dto.Customer.CustomerCreationRequest;
+import com.programmers.springbootbasic.domain.customer.Repository.CustomerRepository;
+import com.programmers.springbootbasic.presentation.controller.dto.Customer.CustomerCreationRequest;
+import com.programmers.springbootbasic.presentation.controller.dto.Customer.CustomerUpdateRequest;
 import com.programmers.springbootbasic.service.dto.Customer.CustomerResponse;
-import com.programmers.springbootbasic.service.dto.Customer.CustomerUpdateRequest;
 import org.springframework.stereotype.Service;
 
 @Service
