+++ src/main/java/com/programmers/springbootbasic/service/VoucherMapper.java
@@ -2,7 +2,7 @@ package com.programmers.springbootbasic.service;
 
 import com.programmers.springbootbasic.domain.voucher.Voucher;
 import com.programmers.springbootbasic.domain.voucher.VoucherType;
-import com.programmers.springbootbasic.service.dto.Voucher.VoucherCreationRequest;
+import com.programmers.springbootbasic.presentation.controller.dto.Voucher.VoucherCreationRequest;
 import com.programmers.springbootbasic.service.dto.Voucher.VoucherResponse;
 import com.programmers.springbootbasic.service.dto.Voucher.VoucherResponses;
