+++ src/test/java/com/programmers/springbootbasic/domain/voucher/PercentDiscountVoucherTest.java
@@ -14,11 +14,10 @@ class PercentDiscountVoucherTest {
     void 정상입력값_바우처생성_성공() {
         // given
         UUID voucherId = UUID.randomUUID();
-        VoucherType voucherType = VoucherType.PERCENT;
         int percent = 30;
 
         // when
-        Voucher percentDiscountVoucher = new PercentDiscountVoucher(voucherId, voucherType, percent);
+        Voucher percentDiscountVoucher = new PercentDiscountVoucher(voucherId, percent);
 
         // then
         assertThat(percentDiscountVoucher).isNotNull();
@@ -28,11 +27,10 @@ class PercentDiscountVoucherTest {
     void 잘못된퍼센트_바우처생성_예외발생() {
         // given
         UUID voucherId = UUID.randomUUID();
-        VoucherType voucherType = VoucherType.PERCENT;
         int percent = 300;
 
         // when && then
-        assertThatThrownBy(() -> new PercentDiscountVoucher(voucherId, voucherType, percent))
+        assertThatThrownBy(() -> new PercentDiscountVoucher(voucherId, percent))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -40,10 +38,8 @@ class PercentDiscountVoucherTest {
     @CsvSource(value = {"4550,33,3050", "5500,10,4950"})
     void 물건금액할인퍼센트_할인_할인된금액(Long price, int percent, Long expectedPrice) {
         // given
-        VoucherType voucherType = VoucherType.PERCENT;
         PercentDiscountVoucher voucher = new PercentDiscountVoucher(
                 UUID.randomUUID(),
-                voucherType,
                 percent
         );
