+++ src/test/java/com/programmers/springbootbasic/domain/voucher/VoucherTest.java
@@ -13,11 +13,10 @@ class VoucherTest {
     void 정상입력값_바우처생성_성공() {
         //given
         UUID voucherId = UUID.randomUUID();
-        VoucherType voucherType = VoucherType.FIX;
         int amount = 10_000;
 
         // when
-        Voucher fixedAmountVoucher = new FixedAmountVoucher(voucherId, voucherType, amount);
+        Voucher fixedAmountVoucher = new FixedAmountVoucher(voucherId, amount);
 
         // then
         assertThat(fixedAmountVoucher).isNotNull();
@@ -27,11 +26,10 @@ class VoucherTest {
     void 잘못된바우처아이디_바우처생성_예외발생() {
         // given
         UUID voucherId = null;
-        VoucherType voucherType = VoucherType.FIX;
         int amount = 10_000;
 
         // when && then
-        assertThatThrownBy(() -> new FixedAmountVoucher(voucherId, voucherType, amount))
+        assertThatThrownBy(() -> new FixedAmountVoucher(voucherId, amount))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
